@@ -5,30 +5,19 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon -e ts --exec \"npm run compile\"",
-    "bridge:start": "nodemon -e ts --exec \"npm run bridge:compile\"",
-    "bridge:compile": "tsc && node ./build/bridge/index.js",
-    "proposal:start": "nodemon -e ts --exec \"npm run proposal:compile\"",
-    "proposal:compile": "tsc && node ./build/proposal/index.js",
-    "dailyLiquidity:start": "nodemon -e ts --exec \"npm run dailyLiquidity:compile\"",
-    "dailyLiquidity:compile": "tsc && node ./build/dailyLiquidity/index.js",
-    "WqtWbnb:start": "nodemon -e ts --exec \"npm run WqtWbnb:compile\"",
-    "WqtWbnb:compile": "tsc && node ./build/WqtWbnb/index.js",
     "compile": "tsc && node ./build/main.js",
     "build": "tsc",
+    "lint": "eslint ./src --fix",
     "test": "lab --sourcemaps --transform node_modules/lab-transform-typescript -v --leaks"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@pancakeswap/sdk": "^2.4.0",
-    "@uniswap/sdk": "^3.0.3",
-    "@uniswap/v2-core": "^1.0.1",
     "@ethersproject/address": "^5.4.0",
     "@ethersproject/contracts": "^5.4.1",
     "@ethersproject/providers": "^5.4.5",
     "@ethersproject/solidity": "^5.4.0",
-    "web3": "^1.5.3",
     "@hapi/basic": "^5.1.1",
     "@hapi/bell": "^12.2.0",
     "@hapi/boom": "^9.1.0",
@@ -37,11 +26,15 @@
     "@hapi/inert": "^6.0.1",
     "@hapi/nes": "^12.0.3",
     "@hapi/vision": "^6.0.0",
+    "@pancakeswap/sdk": "^2.4.0",
     "@types/bluebird": "^3.5.32",
     "@types/dotenv": "^6.1.1",
     "@types/joi": "^14.3.4",
     "@types/validator": "^13.1.0",
-    "@workquest/database-models": "git+https://github.com/WorkQuest/database-models.git#c80b5deb118ba874e60865c9346e1813084945ec",
+    "@uniswap/sdk": "^3.0.3",
+    "@uniswap/v2-core": "^1.0.1",
+    "@workquest/database-models": "git+https://github.com/WorkQuest/database-models.git#cc697155cbaea845c1b8cf8a7bd0aea260ce5556",
+    "amqplib": "^0.8.0",
     "aws-sdk": "^2.918.0",
     "axios": "^0.21.1",
     "bcrypt": "^5.0.0",
@@ -76,26 +69,19 @@
     "speakeasy": "^2.0.0",
     "twilio": "^3.66.0",
     "typescript": "^4.1.5",
-    "uuid": "^8.2.0"
+    "uuid": "^8.2.0",
+    "web3": "^1.5.3"
   },
   "devDependencies": {
     "@hapi/code": "^8.0.3",
-    "@hapi/eslint-config-hapi": "^12.3.0",
-    "@hapi/eslint-plugin-hapi": "^4.3.4",
     "@hapi/lab": "^24.2.1",
     "@types/hapi": "^18.0.3",
     "@types/node": "^12.12.52",
-    "eslint": "^6.5.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.4.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-standard": "^4.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.8.0",
+    "@typescript-eslint/parser": "^5.8.0",
+    "eslint": "^8.7.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "pino": "^6.13.2"
   }
 }
