@@ -1,17 +1,17 @@
-import { getRandomHexToken, output } from "../utils";
-import { addSendEmailJob } from "../jobs/sendEmail";
-import config from "../config/config";
-import * as path from "path";
-import * as fs from "fs";
-import Handlebars = require("handlebars");
-import {
-  User,
-} from "@workquest/database-models/lib/models";
+import { getRandomHexToken, output } from '../utils';
+import { addSendEmailJob } from '../jobs/sendEmail';
+import config from '../config/config';
+import * as path from 'path';
+import * as fs from 'fs';
+import Handlebars = require('handlebars');
+import { User } from '@workquest/database-models/lib/models';
 
-const confirmTemplatePath = path.join(__dirname, "..", "..", "..", "templates", "resetPasswordConfirmation.html");
-const confirmTemplate = Handlebars.compile(fs.readFileSync(confirmTemplatePath, {
-  encoding: "utf-8"
-}));
+const confirmTemplatePath = path.join(__dirname, '..', '..', '..', 'templates', 'resetPasswordConfirmation.html');
+const confirmTemplate = Handlebars.compile(
+  fs.readFileSync(confirmTemplatePath, {
+    encoding: 'utf-8',
+  }),
+);
 
 export async function sendCodeForRestorePassword(r) {
   const user = await User.findWithEmail(r.payload.email);
@@ -23,9 +23,9 @@ export async function sendCodeForRestorePassword(r) {
     const emailHtml = confirmTemplate({ confirmLink: emailRestorePasswordLink, confirmCode: emailRestorePasswordCode });
     await addSendEmailJob({
       email: r.payload.email,
-      subject: "Work Quest | Reset password confirmation",
+      subject: 'Work Quest | Reset password confirmation',
       text: `Change password. Follow this link ${emailRestorePasswordLink}`,
-      html: emailHtml
+      html: emailHtml,
     });
 
     await user.update({ settings: { ...user.settings, restorePassword: emailRestorePasswordCode } });
@@ -35,13 +35,13 @@ export async function sendCodeForRestorePassword(r) {
 }
 
 export async function setNewPassword(r) {
-  const user = await User.scope("withPassword").findOne({
-    where: { "settings.restorePassword": r.payload.token }
+  const user = await User.scope('withPassword').findOne({
+    where: { 'settings.restorePassword': r.payload.token },
   });
 
   if (!user) return output();
 
-  await user.update({ password: r.payload.newPassword , "settings.restorePassword": null });
+  await user.update({ password: r.payload.newPassword, 'settings.restorePassword': null });
 
   return output();
 }
