@@ -1,10 +1,6 @@
-import Joi = require("joi");
-import { getRecipientSwaps } from "../../api/bridge";
-import {
-  limitSchema,
-  offsetSchema,
-  outputOkSchema,
-} from "@workquest/database-models/lib/schemes";
+import Joi = require('joi');
+import { getRecipientSwaps } from '../../api/bridge';
+import { limitSchema, offsetSchema, outputOkSchema } from '@workquest/database-models/lib/schemes';
 
 // TODO: перенести в data-base-models
 const swapSchema = Joi.object({
@@ -24,25 +20,27 @@ const swapSchema = Joi.object({
 
 const swapsSchema = Joi.array().items(swapSchema).label('Swaps');
 
-export default [{
-  method: 'GET',
-  path: '/v1/bridge/recipient/{recipient}/swaps',
-  handler: getRecipientSwaps,
-  options: {
-    auth: false,
-    id: 'v1.bridge.getRecipientSwaps',
-    tags: ['api', 'bridge'],
-    validate: {
-      params: Joi.object({
-        recipient: Joi.string().required().label('RecipientHash'),
-      }).label("GetBridgeSwapsParams"),
-      query: Joi.object({
-        limit: limitSchema,
-        offset: offsetSchema,
-      }).label('GetBridgeSwapsQuery'),
+export default [
+  {
+    method: 'GET',
+    path: '/v1/bridge/recipient/{recipient}/swaps',
+    handler: getRecipientSwaps,
+    options: {
+      auth: false,
+      id: 'v1.bridge.getRecipientSwaps',
+      tags: ['api', 'bridge'],
+      validate: {
+        params: Joi.object({
+          recipient: Joi.string().required().label('RecipientHash'),
+        }).label('GetBridgeSwapsParams'),
+        query: Joi.object({
+          limit: limitSchema,
+          offset: offsetSchema,
+        }).label('GetBridgeSwapsQuery'),
+      },
+      response: {
+        schema: outputOkSchema(swapsSchema).label('UserResponse'),
+      },
     },
-    response: {
-      schema: outputOkSchema(swapsSchema).label("UserResponse")
-    },
-  }
-}];
+  },
+];
