@@ -1,6 +1,6 @@
 import * as main from './server';
 
-try{
+try {
   main.init();
 } catch (e) {
   console.error(e);
