@@ -1,17 +1,17 @@
-import {Op} from "sequelize";
-import {error, output} from "../utils";
-import {Errors} from "../utils/errors";
-import {setMessageAsReadJob} from "../jobs/setMessageAsRead";
-import {updateCountUnreadMessagesJob} from "../jobs/updateCountUnreadMessages";
-import {resetUnreadCountMessagesOfMemberJob} from "../jobs/resetUnreadCountMessagesOfMember";
-import {incrementUnreadCountMessageOfMembersJob} from "../jobs/incrementUnreadCountMessageOfMembers";
-import {updateCountUnreadChatsJob} from "../jobs/updateCountUnreadChats";
-import {ChatController} from "../controllers/chat/controller.chat";
-import {ChatNotificationActions, publishChatNotifications} from "../websocket/websocket.chat";
-import {MediaController} from "../controllers/controller.media";
-import {MessageController} from "../controllers/chat/controller.message";
-import {UserController} from "../controllers/user/controller.user";
-import {listOfUsersByChatsCountQuery, listOfUsersByChatsQuery} from "../queries";
+import { Op } from 'sequelize';
+import { error, output } from '../utils';
+import { Errors } from '../utils/errors';
+import { setMessageAsReadJob } from '../jobs/setMessageAsRead';
+import { updateCountUnreadMessagesJob } from '../jobs/updateCountUnreadMessages';
+import { resetUnreadCountMessagesOfMemberJob } from '../jobs/resetUnreadCountMessagesOfMember';
+import { incrementUnreadCountMessageOfMembersJob } from '../jobs/incrementUnreadCountMessageOfMembers';
+import { updateCountUnreadChatsJob } from '../jobs/updateCountUnreadChats';
+import { ChatController } from '../controllers/chat/controller.chat';
+import { ChatNotificationActions } from '../controllers/controller.broker';
+import { MediaController } from '../controllers/controller.media';
+import { MessageController } from '../controllers/chat/controller.message';
+import { UserController } from '../controllers/user/controller.user';
+import { listOfUsersByChatsCountQuery, listOfUsersByChatsQuery } from '../queries';
 import {
   User,
   Chat,
@@ -25,25 +25,28 @@ import {
   StarredMessage,
   QuestChatStatuses,
   SenderMessageStatus,
-} from "@workquest/database-models/lib/models";
+} from '@workquest/database-models/lib/models';
 
 export async function getUserChats(r) {
-  const include = [{
-    model: ChatMember,
-    where: { userId: r.auth.credentials.id },
-    required: true,
-    as: 'meMember',
-  }, {
-    model: StarredChat,
-    as: 'star',
-    where: { userId: r.auth.credentials.id },
-    required: r.query.starred,
-  }];
+  const include = [
+    {
+      model: ChatMember,
+      where: { userId: r.auth.credentials.id },
+      required: true,
+      as: 'meMember',
+    },
+    {
+      model: StarredChat,
+      as: 'star',
+      where: { userId: r.auth.credentials.id },
+      required: r.query.starred,
+    },
+  ];
 
   const count = await Chat.unscoped().count({ include });
   const chats = await Chat.findAll({
     include,
-    order: [ ['lastMessageDate', r.query.sort.lastMessageDate] ],
+    order: [['lastMessageDate', r.query.sort.lastMessageDate]],
     limit: r.query.limit,
     offset: r.query.offset,
   });
@@ -58,7 +61,7 @@ export async function getChatMessages(r) {
       where: { userId: r.auth.credentials.id },
       required: false,
       as: 'meMember',
-    }
+    },
   });
   const chatController = new ChatController(chat);
 
@@ -66,30 +69,32 @@ export async function getChatMessages(r) {
 
   const { count, rows } = await Message.findAndCountAll({
     where: { chatId: chat.id },
-    include: [{
-      model: StarredMessage,
-      as: "star",
-      where: { userId: r.auth.credentials.id },
-      required: r.query.starred,
-    }],
+    include: [
+      {
+        model: StarredMessage,
+        as: 'star',
+        where: { userId: r.auth.credentials.id },
+        required: r.query.starred,
+      },
+    ],
     distinct: true,
     limit: r.query.limit,
     offset: r.query.offset,
-    order: [ ['createdAt', r.query.sort.createdAt] ],
+    order: [['createdAt', r.query.sort.createdAt]],
   });
 
   return output({ count, messages: rows, chat });
 }
 
 export async function getUserChat(r) {
-  const chat = await Chat.findByPk(r.params.chatId,{
+  const chat = await Chat.findByPk(r.params.chatId, {
     include: {
       model: StarredChat,
-      as: "star",
+      as: 'star',
       required: false,
-    }
+    },
   });
-  const chatController = new ChatController(chat)
+  const chatController = new ChatController(chat);
 
   await chatController.chatMustHaveMember(r.auth.credentials.id);
 
@@ -103,23 +108,23 @@ export async function listOfUsersByChats(r) {
       limitValue: r.query.limit,
       offsetValue: r.query.offset,
       excludeUsersFromChatId: r.query.excludeMembersChatId || '',
-    }
-  }
+    },
+  };
 
-  const [countResults, ] = await r.server.app.db.query(listOfUsersByChatsCountQuery, options);
-  const [userResults, ] = await r.server.app.db.query(listOfUsersByChatsQuery, options);
+  const [countResults] = await r.server.app.db.query(listOfUsersByChatsCountQuery, options);
+  const [userResults] = await r.server.app.db.query(listOfUsersByChatsQuery, options);
 
-  const users = userResults.map(result => ({
+  const users = userResults.map((result) => ({
     id: result.id,
     firstName: result.firstName,
     lastName: result.lastName,
     additionalInfo: result.additionalInfo,
     avatarId: result.avatarId,
     avatar: {
-      id: result["avatar.id"],
-      url: result["avatar.url"],
-      contentType: result["avatar.contentType"],
-    }
+      id: result['avatar.id'],
+      url: result['avatar.url'],
+      contentType: result['avatar.contentType'],
+    },
   }));
 
   return output({ count: parseInt(countResults[0].count), users });
@@ -132,12 +137,14 @@ export async function getChatMembers(r) {
   await chatController.chatMustHaveMember(r.auth.credentials.id);
 
   const { count, rows } = await User.scope('shortWithAdditionalInfo').findAndCountAll({
-    include: [{
-      model: ChatMember,
-      attributes: [],
-      as: 'chatMember',
-      where: { chatId: chat.id }
-    }],
+    include: [
+      {
+        model: ChatMember,
+        attributes: [],
+        as: 'chatMember',
+        where: { chatId: chat.id },
+      },
+    ],
     limit: r.query.limit,
     offset: r.query.offset,
   });
@@ -156,17 +163,20 @@ export async function createGroupChat(r) {
 
   const transaction = await r.server.app.db.transaction();
 
-  const groupChat = await Chat.create({
-    name: r.payload.name,
-    ownerUserId: r.auth.credentials.id,
-    type: ChatType.group,
-  }, { transaction });
+  const groupChat = await Chat.create(
+    {
+      name: r.payload.name,
+      ownerUserId: r.auth.credentials.id,
+      type: ChatType.group,
+    },
+    { transaction },
+  );
 
   const message = await Message.build({
     senderUserId: r.auth.credentials.id,
     chatId: groupChat.id,
     type: MessageType.info,
-    number: 1, /** Because created */
+    number: 1 /** Because created */,
   });
 
   const infoMessage = await InfoMessage.build({
@@ -174,24 +184,24 @@ export async function createGroupChat(r) {
     messageAction: MessageAction.groupChatCreate,
   });
 
-  await Promise.all([
-    message.save({ transaction }),
-    infoMessage.save({ transaction }),
-  ]);
+  await Promise.all([message.save({ transaction }), infoMessage.save({ transaction })]);
 
-  await groupChat.update({
-    lastMessageId: message.id,
-    lastMessageDate: message.createdAt,
-  }, { transaction })
+  await groupChat.update(
+    {
+      lastMessageId: message.id,
+      lastMessageDate: message.createdAt,
+    },
+    { transaction },
+  );
 
-  const chatMembers = memberUserIds.map(userId => {
+  const chatMembers = memberUserIds.map((userId) => {
     return {
       userId,
       chatId: groupChat.id,
-      unreadCountMessages: (userId === r.auth.credentials.id ? 0 : 1),
-      lastReadMessageId: (userId === r.auth.credentials.id ? message.id : null),
-      lastReadMessageNumber: (userId === r.auth.credentials.id ? message.number : null),
-    }
+      unreadCountMessages: userId === r.auth.credentials.id ? 0 : 1,
+      lastReadMessageId: userId === r.auth.credentials.id ? message.id : null,
+      lastReadMessageNumber: userId === r.auth.credentials.id ? message.number : null,
+    };
   });
 
   await ChatMember.bulkCreate(chatMembers, { transaction });
@@ -202,8 +212,8 @@ export async function createGroupChat(r) {
 
   await updateCountUnreadChatsJob({ userIds: memberUserIds });
 
-  await publishChatNotifications(r.server, {
-    recipients: memberUserIds.filter(userId => userId !== r.auth.credentials.id),
+  r.server.app.broker.sendChatNotification({
+    recipients: memberUserIds.filter((userId) => userId !== r.auth.credentials.id),
     action: ChatNotificationActions.groupChatCreate,
     data: result, // TODO lastReadMessageId: message.id
   });
@@ -231,56 +241,69 @@ export async function sendMessageToUser(r) {
 
   const [chat, isChatCreated] = await Chat.findOrCreate({
     where: { type: ChatType.private },
-    include: [{
-      model: ChatMember,
-      as: 'firstMemberInPrivateChat',
-      where: { userId: r.params.userId },
-      required: true,
-      attributes: [],
-    }, {
-      model: ChatMember,
-      as: 'secondMemberInPrivateChat',
-      where: { userId: r.auth.credentials.id },
-      required: true,
-      attributes: [],
-    }],
+    include: [
+      {
+        model: ChatMember,
+        as: 'firstMemberInPrivateChat',
+        where: { userId: r.params.userId },
+        required: true,
+        attributes: [],
+      },
+      {
+        model: ChatMember,
+        as: 'secondMemberInPrivateChat',
+        where: { userId: r.auth.credentials.id },
+        required: true,
+        attributes: [],
+      },
+    ],
     defaults: {
       type: ChatType.private,
       lastMessageId: message.id,
       lastMessageDate: message.createdAt,
-    }, transaction,
+    },
+    transaction,
   });
 
   const lastMessage = await Message.findOne({
-    order: [ ['createdAt', 'DESC'] ],
-    where: { chatId: chat.id }
+    order: [['createdAt', 'DESC']],
+    where: { chatId: chat.id },
   });
 
   message.chatId = chat.id;
-  message.number = lastMessage ? message.number = lastMessage.number + 1 : message.number = 1;
+  message.number = lastMessage ? (message.number = lastMessage.number + 1) : (message.number = 1);
 
   await message.save({ transaction });
   await message.$set('medias', medias, { transaction });
 
   if (isChatCreated) {
-    await ChatMember.bulkCreate([{
-      unreadCountMessages: 0, /** Because created */
-      chatId: chat.id,
-      userId: r.auth.credentials.id,
-      lastReadMessageId: message.id, /** Because created */
-      lastReadMessageNumber: message.number,
-    }, {
-      unreadCountMessages: 1, /** Because created */
-      chatId: chat.id,
-      userId: r.params.userId,
-      lastReadMessageId: null, /** Because created */
-      lastReadMessageNumber: null,
-    }], { transaction })
+    await ChatMember.bulkCreate(
+      [
+        {
+          unreadCountMessages: 0 /** Because created */,
+          chatId: chat.id,
+          userId: r.auth.credentials.id,
+          lastReadMessageId: message.id /** Because created */,
+          lastReadMessageNumber: message.number,
+        },
+        {
+          unreadCountMessages: 1 /** Because created */,
+          chatId: chat.id,
+          userId: r.params.userId,
+          lastReadMessageId: null /** Because created */,
+          lastReadMessageNumber: null,
+        },
+      ],
+      { transaction },
+    );
   } else {
-    await chat.update({
-      lastMessageId: message.id,
-      lastMessageDate: message.createdAt,
-    }, { transaction });
+    await chat.update(
+      {
+        lastMessageId: message.id,
+        lastMessageDate: message.createdAt,
+      },
+      { transaction },
+    );
   }
 
   await transaction.commit();
@@ -294,7 +317,8 @@ export async function sendMessageToUser(r) {
     });
 
     await incrementUnreadCountMessageOfMembersJob({
-      chatId: chat.id, notifierUserId: r.auth.credentials.id,
+      chatId: chat.id,
+      notifierUserId: r.auth.credentials.id,
     });
   }
 
@@ -309,7 +333,7 @@ export async function sendMessageToUser(r) {
 
   const result = await Message.findByPk(message.id);
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.newMessage,
     recipients: [r.params.userId],
     data: result,
@@ -332,32 +356,38 @@ export async function sendMessageToChat(r) {
   const transaction = await r.server.app.db.transaction();
 
   const lastMessage = await Message.findOne({
-    order: [ ['createdAt', 'DESC'] ],
-    where: { chatId: chat.id }
+    order: [['createdAt', 'DESC']],
+    where: { chatId: chat.id },
   });
 
-  const message = await Message.create({
-    senderUserId: r.auth.credentials.id,
-    chatId: chat.id,
-    type: MessageType.message,
-    text: r.payload.text,
-    senderStatus: SenderMessageStatus.unread,
-    number: lastMessage.number + 1,
-  }, { transaction });
+  const message = await Message.create(
+    {
+      senderUserId: r.auth.credentials.id,
+      chatId: chat.id,
+      type: MessageType.message,
+      text: r.payload.text,
+      senderStatus: SenderMessageStatus.unread,
+      number: lastMessage.number + 1,
+    },
+    { transaction },
+  );
 
   await message.$set('medias', medias, { transaction });
 
-  await chat.update({
-    lastMessageId: message.id,
-    lastMessageDate: message.createdAt,
-  }, { transaction });
+  await chat.update(
+    {
+      lastMessageId: message.id,
+      lastMessageDate: message.createdAt,
+    },
+    { transaction },
+  );
 
   await transaction.commit();
 
   const membersWithoutSender = await ChatMember.scope('userIdsOnly').findAll({
-    where: { chatId: chat.id, userId: { [Op.ne]: r.auth.credentials.id } }
+    where: { chatId: chat.id, userId: { [Op.ne]: r.auth.credentials.id } },
   });
-  const userIdsWithoutSender = membersWithoutSender.map(member => member.userId);
+  const userIdsWithoutSender = membersWithoutSender.map((member) => member.userId);
   const result = await Message.findByPk(message.id);
 
   await resetUnreadCountMessagesOfMemberJob({
@@ -368,7 +398,8 @@ export async function sendMessageToChat(r) {
   });
 
   await incrementUnreadCountMessageOfMembersJob({
-    chatId: chat.id, notifierUserId: r.auth.credentials.id,
+    chatId: chat.id,
+    notifierUserId: r.auth.credentials.id,
   });
 
   await setMessageAsReadJob({
@@ -378,10 +409,10 @@ export async function sendMessageToChat(r) {
   });
 
   await updateCountUnreadChatsJob({
-    userIds: [r.auth.credentials.id, ...userIdsWithoutSender]
+    userIds: [r.auth.credentials.id, ...userIdsWithoutSender],
   });
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.newMessage,
     recipients: userIdsWithoutSender,
     data: result,
@@ -397,10 +428,7 @@ export async function addUsersInGroupChat(r) {
   const groupChat = await Chat.findByPk(r.params.chatId);
   const chatController = new ChatController(groupChat);
 
-  await chatController
-    .chatMustHaveType(ChatType.group)
-    .chatMustHaveOwner(r.auth.credentials.id)
-    .usersNotExistInGroupChat(userIds)
+  await chatController.chatMustHaveType(ChatType.group).chatMustHaveOwner(r.auth.credentials.id).usersNotExistInGroupChat(userIds);
 
   const messages: Message[] = [];
   const infoMessages: InfoMessage[] = [];
@@ -429,36 +457,39 @@ export async function addUsersInGroupChat(r) {
   const lastMessage = messages[messages.length - 1];
   const transaction = await r.server.app.db.transaction();
 
-  const members = userIds.map(userId => {
+  const members = userIds.map((userId) => {
     return {
       chatId: groupChat.id,
       userId: userId,
-      unreadCountMessages: 1, /** Because info message */
-      lastReadMessageId: groupChat.lastMessage.id, /** Because new member */
-    }
+      unreadCountMessages: 1 /** Because info message */,
+      lastReadMessageId: groupChat.lastMessage.id /** Because new member */,
+    };
   });
   await ChatMember.bulkCreate(members, { transaction });
 
-  let messagesResult = await Promise.all(messages.map(_ => _.save({ transaction })));
-  const infoMessagesResult = await Promise.all(infoMessages.map(_ => _.save({ transaction })));
+  let messagesResult = await Promise.all(messages.map((_) => _.save({ transaction })));
+  const infoMessagesResult = await Promise.all(infoMessages.map((_) => _.save({ transaction })));
 
-  await groupChat.update({
-    lastMessageId: lastMessage.id,
-    lastMessageDate: lastMessage.createdAt,
-  }, { transaction });
+  await groupChat.update(
+    {
+      lastMessageId: lastMessage.id,
+      lastMessageDate: lastMessage.createdAt,
+    },
+    { transaction },
+  );
 
   await transaction.commit();
 
   const membersWithoutSender = await ChatMember.scope('userIdsOnly').findAll({
-    where: { chatId: groupChat.id, userId: { [Op.ne]: r.auth.credentials.id } }
+    where: { chatId: groupChat.id, userId: { [Op.ne]: r.auth.credentials.id } },
   });
-  const userIdsInChatWithoutSender = membersWithoutSender.map(member => member.userId);
+  const userIdsInChatWithoutSender = membersWithoutSender.map((member) => member.userId);
 
-  messagesResult = messagesResult.map(message => {
+  messagesResult = messagesResult.map((message) => {
     const keysMessage: { [key: string]: any } = message.toJSON();
-    const keysInfoMessage = infoMessagesResult.find(_ => _.messageId === message.id).toJSON() as InfoMessage;
+    const keysInfoMessage = infoMessagesResult.find((_) => _.messageId === message.id).toJSON() as InfoMessage;
 
-    keysInfoMessage.user = users.find(_ => _.id === keysInfoMessage.userId).toJSON() as User;
+    keysInfoMessage.user = users.find((_) => _.id === keysInfoMessage.userId).toJSON() as User;
 
     keysMessage.infoMessage = keysInfoMessage;
 
@@ -478,10 +509,10 @@ export async function addUsersInGroupChat(r) {
   });
 
   await updateCountUnreadChatsJob({
-    userIds: [r.auth.credentials.id, ...userIdsInChatWithoutSender]
+    userIds: [r.auth.credentials.id, ...userIdsInChatWithoutSender],
   });
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.groupChatAddUser,
     recipients: userIdsInChatWithoutSender,
     data: messagesResult,
@@ -496,9 +527,7 @@ export async function removeUserInGroupChat(r) {
   const groupChat = await Chat.findByPk(r.params.chatId);
   const chatController = new ChatController(groupChat);
 
-  chatController
-    .chatMustHaveType(ChatType.group)
-    .chatMustHaveOwner(r.auth.credentials.id)
+  chatController.chatMustHaveType(ChatType.group).chatMustHaveOwner(r.auth.credentials.id);
 
   await chatController.chatMustHaveMember(r.params.userId);
 
@@ -508,34 +537,44 @@ export async function removeUserInGroupChat(r) {
     where: {
       chatId: groupChat.id,
       userId: r.params.userId,
-    }, transaction,
+    },
+    transaction,
   });
 
-  const message = await Message.create({
-    senderUserId: r.auth.credentials.id,
-    chatId: groupChat.id,
-    type: MessageType.info,
-    number: groupChat.lastMessage.number + 1
-  }, { transaction });
+  const message = await Message.create(
+    {
+      senderUserId: r.auth.credentials.id,
+      chatId: groupChat.id,
+      type: MessageType.info,
+      number: groupChat.lastMessage.number + 1,
+    },
+    { transaction },
+  );
 
-  await InfoMessage.create({
-    userId: r.params.userId,
-    messageId: message.id,
-    messageAction: MessageAction.groupChatDeleteUser,
-  }, { transaction });
+  await InfoMessage.create(
+    {
+      userId: r.params.userId,
+      messageId: message.id,
+      messageAction: MessageAction.groupChatDeleteUser,
+    },
+    { transaction },
+  );
 
-  await groupChat.update({
-    lastMessageId: message.id,
-    lastMessageDate: message.createdAt,
-  }, { transaction });
+  await groupChat.update(
+    {
+      lastMessageId: message.id,
+      lastMessageDate: message.createdAt,
+    },
+    { transaction },
+  );
 
   await transaction.commit();
 
   const result = await Message.findByPk(message.id);
   const membersWithoutSender = await ChatMember.scope('userIdsOnly').findAll({
-    where: { chatId: groupChat.id, userId: { [Op.ne]: r.auth.credentials.id } }
+    where: { chatId: groupChat.id, userId: { [Op.ne]: r.auth.credentials.id } },
   });
-  const userIdsWithoutSender = membersWithoutSender.map(member => member.userId);
+  const userIdsWithoutSender = membersWithoutSender.map((member) => member.userId);
 
   await resetUnreadCountMessagesOfMemberJob({
     chatId: groupChat.id,
@@ -550,10 +589,10 @@ export async function removeUserInGroupChat(r) {
   });
 
   await updateCountUnreadChatsJob({
-    userIds: [r.auth.credentials.id, ...userIdsWithoutSender]
+    userIds: [r.auth.credentials.id, ...userIdsWithoutSender],
   });
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.groupChatDeleteUser,
     recipients: userIdsWithoutSender,
     data: result,
@@ -570,41 +609,50 @@ export async function leaveFromGroupChat(r) {
   await chatController.chatMustHaveMember(r.auth.credentials.id);
 
   if (groupChat.ownerUserId === r.auth.credentials.id) {
-    return error(Errors.Forbidden, "User is chat owner", {}); // TODO
+    return error(Errors.Forbidden, 'User is chat owner', {}); // TODO
   }
 
   const transaction = await r.server.app.db.transaction();
 
   await ChatMember.destroy({
     where: { chatId: groupChat.id, userId: r.auth.credentials.id },
-    transaction
+    transaction,
   });
 
-  const message = await Message.create({
-    senderUserId: r.auth.credentials.id,
-    chatId: groupChat.id,
-    type: MessageType.info,
-    number: groupChat.lastMessage.number + 1,
-  }, { transaction });
+  const message = await Message.create(
+    {
+      senderUserId: r.auth.credentials.id,
+      chatId: groupChat.id,
+      type: MessageType.info,
+      number: groupChat.lastMessage.number + 1,
+    },
+    { transaction },
+  );
 
-  await InfoMessage.create({
-    messageId: message.id,
-    // userId: r.auth.credentials.id,
-    messageAction: MessageAction.groupChatLeaveUser,
-  }, { transaction });
+  await InfoMessage.create(
+    {
+      messageId: message.id,
+      // userId: r.auth.credentials.id,
+      messageAction: MessageAction.groupChatLeaveUser,
+    },
+    { transaction },
+  );
 
-  await groupChat.update({
-    lastMessageId: message.id,
-    lastMessageDate: message.createdAt,
-  }, { transaction });
+  await groupChat.update(
+    {
+      lastMessageId: message.id,
+      lastMessageDate: message.createdAt,
+    },
+    { transaction },
+  );
 
   await transaction.commit();
 
   const result = await Message.findByPk(message.id);
   const membersWithoutSender = await ChatMember.scope('userIdsOnly').findAll({
-    where: { chatId: groupChat.id, userId: { [Op.ne]: r.auth.credentials.id } }
+    where: { chatId: groupChat.id, userId: { [Op.ne]: r.auth.credentials.id } },
   });
-  const userIdsWithoutSender = membersWithoutSender.map(member => member.userId);
+  const userIdsWithoutSender = membersWithoutSender.map((member) => member.userId);
 
   await incrementUnreadCountMessageOfMembersJob({
     chatId: groupChat.id,
@@ -612,10 +660,10 @@ export async function leaveFromGroupChat(r) {
   });
 
   await updateCountUnreadChatsJob({
-    userIds: [r.auth.credentials.id, ...userIdsWithoutSender]
+    userIds: [r.auth.credentials.id, ...userIdsWithoutSender],
   });
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.groupChatLeaveUser,
     recipients: userIdsWithoutSender,
     data: result,
@@ -633,18 +681,18 @@ export async function setMessagesAsRead(r) {
   const message = await Message.findByPk(r.payload.messageId);
 
   if (!message) {
-    return error(Errors.NotFound, "Message is not found", {});
+    return error(Errors.NotFound, 'Message is not found', {});
   }
 
   const otherSenders = await Message.unscoped().findAll({
-    attributes: ["senderUserId"],
+    attributes: ['senderUserId'],
     where: {
       chatId: chatController.chat.id,
       senderUserId: { [Op.ne]: r.auth.credentials.id },
       senderStatus: SenderMessageStatus.unread,
       number: { [Op.gte]: message.number },
     },
-    group: ["senderUserId"]
+    group: ['senderUserId'],
   });
 
   await updateCountUnreadMessagesJob({
@@ -660,16 +708,16 @@ export async function setMessagesAsRead(r) {
   await setMessageAsReadJob({
     lastUnreadMessage: { id: message.id, number: message.number },
     chatId: r.params.chatId,
-    senderId: r.auth.credentials.id
+    senderId: r.auth.credentials.id,
   });
 
   await updateCountUnreadChatsJob({
-    userIds: [r.auth.credentials.id]
+    userIds: [r.auth.credentials.id],
   });
 
-  await publishChatNotifications(r.server, {
+  r.server.app.broker.sendChatNotification({
     action: ChatNotificationActions.messageReadByRecipient,
-    recipients: otherSenders.map(sender => sender.senderUserId),
+    recipients: otherSenders.map((sender) => sender.senderUserId),
     data: message,
   });
 
@@ -681,15 +729,18 @@ export async function getUserStarredMessages(r) {
     distinct: true,
     limit: r.query.limit,
     offset: r.query.offset,
-    include: [{
-      model: StarredMessage,
-      as: "star",
-      where: { userId: r.auth.credentials.id },
-      required: true,
-    }, {
-      model: Chat.unscoped(),
-      as: "chat",
-    }]
+    include: [
+      {
+        model: StarredMessage,
+        as: 'star',
+        where: { userId: r.auth.credentials.id },
+        required: true,
+      },
+      {
+        model: Chat.unscoped(),
+        as: 'chat',
+      },
+    ],
   });
 
   return output({ count, messages: rows });
@@ -716,8 +767,8 @@ export async function removeStarFromMessage(r) {
   const starredMessage = await StarredMessage.findOne({
     where: {
       messageId: r.params.messageId,
-      userId: r.auth.credentials.id
-    }
+      userId: r.auth.credentials.id,
+    },
   });
 
   if (!starredMessage) {
@@ -752,8 +803,8 @@ export async function removeStarFromChat(r) {
   await StarredChat.destroy({
     where: {
       chatId: r.params.chatId,
-      userId: r.auth.credentials.id
-    }
+      userId: r.auth.credentials.id,
+    },
   });
 
   return output();
