@@ -1,77 +1,76 @@
-import { Op } from "sequelize";
-import { error, output } from "../utils";
-import { Errors } from "../utils/errors";
-import { MediaController } from "../controllers/controller.media";
-import {
-  User,
-  Discussion,
-  DiscussionLike,
-  DiscussionComment,
-  DiscussionCommentLike,
-  StarredDiscussion
-} from "@workquest/database-models/lib/models";
+import { Op } from 'sequelize';
+import { error, output } from '../utils';
+import { Errors } from '../utils/errors';
+import { MediaController } from '../controllers/controller.media';
+import { User, Discussion, DiscussionLike, DiscussionComment, DiscussionCommentLike, StarredDiscussion } from '@workquest/database-models/lib/models';
 
-const searchFields = [
-  'title',
-  'description',
-];
+const searchFields = ['title', 'description'];
 
 export async function getDiscussion(r) {
   const discussion = await Discussion.findOne({
-    include: [{
-      model: DiscussionLike,
-      as: "liked",
-      where: { userId: r.auth.credentials.id },
-      required: false
-    }, {
-      model: StarredDiscussion,
-      as: 'star',
-      where: { userId: r.auth.credentials.id },
-      required: false
-    }, {
-      model: User.scope('short'),
-      as: "author",
-    }],
-    where: { id: r.params.discussionId }
+    include: [
+      {
+        model: DiscussionLike,
+        as: 'liked',
+        where: { userId: r.auth.credentials.id },
+        required: false,
+      },
+      {
+        model: StarredDiscussion,
+        as: 'star',
+        where: { userId: r.auth.credentials.id },
+        required: false,
+      },
+      {
+        model: User.scope('short'),
+        as: 'author',
+      },
+    ],
+    where: { id: r.params.discussionId },
   });
 
   if (!discussion) {
-    return error(Errors.NotFound, "Discussion not found", { discussionId: r.params.discussionId });
+    return error(Errors.NotFound, 'Discussion not found', { discussionId: r.params.discussionId });
   }
 
   return output(discussion);
 }
 
 export async function getDiscussions(r) {
-  const where = { };
+  const where = {};
 
-  const include = [{
-    model: DiscussionLike,
-    as: "liked",
-    where: { userId: r.auth.credentials.id },
-    required: false,
-  }, {
-    model: StarredDiscussion,
-    as: 'star',
-    where: { userId: r.auth.credentials.id },
-    required: r.query.starred,
-  }, {
-    model: User,
-    as: "author",
-  }];
+  const include = [
+    {
+      model: DiscussionLike,
+      as: 'liked',
+      where: { userId: r.auth.credentials.id },
+      required: false,
+    },
+    {
+      model: StarredDiscussion,
+      as: 'star',
+      where: { userId: r.auth.credentials.id },
+      required: r.query.starred,
+    },
+    {
+      model: User,
+      as: 'author',
+    },
+  ];
 
   if (r.query.q) {
-    where[Op.or] = searchFields.map(field => ({
-      [field]: { [Op.iLike]: `%${r.query.q}%` }
+    where[Op.or] = searchFields.map((field) => ({
+      [field]: { [Op.iLike]: `%${r.query.q}%` },
     }));
   }
 
   const { count, rows } = await Discussion.findAndCountAll({
     distinct: true,
-    include, where,
-    order: [["createdAt", "DESC"]],
+    include,
+    where,
+    order: [['createdAt', 'DESC']],
     limit: r.query.limit,
-    offset: r.query.offset
+    offset: r.query.offset,
   });
 
   return output({ count, discussions: rows });
@@ -81,20 +80,22 @@ export async function getSubComments(r) {
   const rootComment = await DiscussionComment.findByPk(r.params.commentId);
 
   if (!rootComment) {
-    return error(Errors.NotFound, "Root comment not found", {});
+    return error(Errors.NotFound, 'Root comment not found', {});
   }
 
   const { count, rows } = await DiscussionComment.findAndCountAll({
-    include: [{
-      model: DiscussionCommentLike,
-      as: "commentLikes",
-      where: { userId: r.auth.credentials.id },
-      required: false
-    }],
+    include: [
+      {
+        model: DiscussionCommentLike,
+        as: 'commentLikes',
+        where: { userId: r.auth.credentials.id },
+        required: false,
+      },
+    ],
     where: { rootCommentId: rootComment.id },
-    order: [["createdAt", "DESC"]],
+    order: [['createdAt', 'DESC']],
     limit: r.query.limit,
-    offset: r.query.offset
+    offset: r.query.offset,
   });
 
   return output({ count, comments: rows });
@@ -104,20 +105,22 @@ export async function getRootComments(r) {
   const discussion = await Discussion.findByPk(r.params.discussionId);
 
   if (!discussion) {
-    return error(Errors.NotFound, "Discussion not found", {});
+    return error(Errors.NotFound, 'Discussion not found', {});
   }
 
   const { count, rows } = await DiscussionComment.findAndCountAll({
-    include: [{
-      model: DiscussionCommentLike,
-      as: "commentLikes",
-      where: { userId: r.auth.credentials.id },
-      required: false
-    }],
+    include: [
+      {
+        model: DiscussionCommentLike,
+        as: 'commentLikes',
+        where: { userId: r.auth.credentials.id },
+        required: false,
+      },
+    ],
     where: { rootCommentId: null, discussionId: r.params.discussionId },
-    order: [["createdAt", "DESC"]],
+    order: [['createdAt', 'DESC']],
     limit: r.query.limit,
-    offset: r.query.offset
+    offset: r.query.offset,
   });
 
   return output({ count, comments: rows });
@@ -127,13 +130,16 @@ export async function createDiscussion(r) {
   const medias = await MediaController.getMedias(r.payload.medias);
   const transaction = await r.server.app.db.transaction();
 
-  const discussion = await Discussion.create({
-    authorId: r.auth.credentials.id,
-    title: r.payload.title,
-    description: r.payload.description
-  }, { transaction });
+  const discussion = await Discussion.create(
+    {
+      authorId: r.auth.credentials.id,
+      title: r.payload.title,
+      description: r.payload.description,
+    },
+    { transaction },
+  );
 
-  await discussion.$set("medias", medias, { transaction });
+  await discussion.$set('medias', medias, { transaction });
 
   await transaction.commit();
 
@@ -144,42 +150,46 @@ export async function sendComment(r) {
   const medias = await MediaController.getMedias(r.payload.medias);
 
   const discussion = await Discussion.findOne({
-    where: { id: r.params.discussionId }
+    where: { id: r.params.discussionId },
   });
 
   if (!discussion) {
-    return error(Errors.NotFound, "Discussion not found", {});
+    return error(Errors.NotFound, 'Discussion not found', {});
   }
 
   let commentLevel = 0;
 
   const transaction = await r.server.app.db.transaction();
 
-  if (r.payload.rootCommentId) { // TODO в джобу
+  if (r.payload.rootCommentId) {
+    // TODO в джобу
     const rootComment = await DiscussionComment.findByPk(r.payload.rootCommentId);
 
     if (!rootComment) {
       await transaction.rollback();
 
-      return error(Errors.NotFound, "Discussion comment not found", {});
+      return error(Errors.NotFound, 'Discussion comment not found', {});
     }
 
-    await rootComment.increment("amountSubComments", { transaction });
+    await rootComment.increment('amountSubComments', { transaction });
 
-    commentLevel = rootComment.level + 1
+    commentLevel = rootComment.level + 1;
   } else {
-    await discussion.increment("amountComments", { transaction });
+    await discussion.increment('amountComments', { transaction });
   }
 
-  const comment = await DiscussionComment.create({
-    authorId: r.auth.credentials.id,
-    discussionId: r.params.discussionId,
-    rootCommentId: r.payload.rootCommentId,
-    text: r.payload.text,
-    level: commentLevel,
-  }, { transaction });
+  const comment = await DiscussionComment.create(
+    {
+      authorId: r.auth.credentials.id,
+      discussionId: r.params.discussionId,
+      rootCommentId: r.payload.rootCommentId,
+      text: r.payload.text,
+      level: commentLevel,
+    },
+    { transaction },
+  );
 
-  await comment.$set("medias", medias, { transaction });
+  await comment.$set('medias', medias, { transaction });
 
   await transaction.commit();
 
@@ -190,7 +200,7 @@ export async function putDiscussionLike(r) {
   const discussion = await Discussion.findByPk(r.params.discussionId);
 
   if (!discussion) {
-    return error(Errors.NotFound, "Discussion not found", {});
+    return error(Errors.NotFound, 'Discussion not found', {});
   }
 
   const transaction = await r.server.app.db.transaction();
@@ -198,11 +208,11 @@ export async function putDiscussionLike(r) {
   const [like, isCreated] = await DiscussionLike.findOrCreate({
     where: { userId: r.auth.credentials.id, discussionId: r.params.discussionId },
     defaults: { userId: r.auth.credentials.id, discussionId: r.params.discussionId },
-    transaction
+    transaction,
   });
 
   if (isCreated) {
-    await discussion.increment("amountLikes", { transaction });
+    await discussion.increment('amountLikes', { transaction });
   }
   await transaction.commit();
 
@@ -213,17 +223,17 @@ export async function removeDiscussionLike(r) {
   const discussion = await Discussion.findByPk(r.params.discussionId);
 
   if (!discussion) {
-    return error(Errors.NotFound, "Discussion not found", {});
+    return error(Errors.NotFound, 'Discussion not found', {});
   }
 
   const transaction = await r.server.app.db.transaction();
 
   await DiscussionLike.destroy({
     where: { discussionId: r.params.discussionId, userId: r.auth.credentials.id },
-    transaction
+    transaction,
   });
 
-  await discussion.decrement("amountLikes", { transaction });
+  await discussion.decrement('amountLikes', { transaction });
 
   await transaction.commit();
 
@@ -234,18 +244,18 @@ export async function putCommentLike(r) {
   const comment = await DiscussionComment.findByPk(r.params.commentId);
 
   if (!comment) {
-    return error(Errors.NotFound, "Comment not found", {});
+    return error(Errors.NotFound, 'Comment not found', {});
   }
 
   const transaction = await r.server.app.db.transaction();
 
   const [like, isCreated] = await DiscussionCommentLike.findOrCreate({
     where: { userId: r.auth.credentials.id, commentId: r.params.commentId },
-    defaults: { userId: r.auth.credentials.id, commentId: r.params.commentId }
+    defaults: { userId: r.auth.credentials.id, commentId: r.params.commentId },
   });
 
   if (isCreated) {
-    await comment.increment("amountLikes", { transaction });
+    await comment.increment('amountLikes', { transaction });
   }
 
   await transaction.commit();
@@ -257,16 +267,16 @@ export async function removeCommentLike(r) {
   const comment = await DiscussionComment.findByPk(r.params.commentId);
 
   if (!comment) {
-    return error(Errors.NotFound, "Comment not found", {});
+    return error(Errors.NotFound, 'Comment not found', {});
   }
 
   const transaction = await r.server.app.db.transaction();
 
-  await comment.decrement("amountLikes", { transaction });
+  await comment.decrement('amountLikes', { transaction });
 
   await DiscussionCommentLike.destroy({
     where: { commentId: r.params.commentId, userId: r.auth.credentials.id },
-    transaction
+    transaction,
   });
 
   await transaction.commit();
@@ -275,26 +285,30 @@ export async function removeCommentLike(r) {
 }
 
 export async function getDiscussionUsersLikes(r) {
-  const { count, rows } = await User.scope("short").findAndCountAll({
-    include: [{
-      model: DiscussionLike,
-      as: "discussionLikes",
-      where: { discussionId: r.params.discussionId }
-    }],
+  const { count, rows } = await User.scope('short').findAndCountAll({
+    include: [
+      {
+        model: DiscussionLike,
+        as: 'discussionLikes',
+        where: { discussionId: r.params.discussionId },
+      },
+    ],
     limit: r.query.limit,
-    offset: r.query.offset
+    offset: r.query.offset,
   });
 
   return output({ count, users: rows });
 }
 
 export async function getCommentUsersLikes(r) {
-  const { count, rows } = await User.scope("short").findAndCountAll({
-    include: [{
-      model: DiscussionCommentLike,
-      as: "commentLikes",
-      where: { commentId: r.params.commentId }
-    }],
+  const { count, rows } = await User.scope('short').findAndCountAll({
+    include: [
+      {
+        model: DiscussionCommentLike,
+        as: 'commentLikes',
+        where: { commentId: r.params.commentId },
+      },
+    ],
     limit: r.query.limit,
     offset: r.query.offset,
   });
@@ -306,19 +320,19 @@ export async function markDiscussionStar(r) {
   const discussion = await Discussion.findByPk(r.params.discussionId);
 
   if (!discussion) {
-    return error(Errors.NotFound, "Discussion is not found", {});
+    return error(Errors.NotFound, 'Discussion is not found', {});
   }
 
   const [star, isCreated] = await StarredDiscussion.findOrCreate({
     where: { discussionId: r.params.discussionId, userId: r.auth.credentials.id },
     defaults: {
       userId: r.auth.credentials.id,
-      discussionId: r.params.discussionId
-    }
+      discussionId: r.params.discussionId,
+    },
   });
 
   if (!isCreated) {
-    return error(Errors.AlreadyExists, "You already set star on this discussion", {});
+    return error(Errors.AlreadyExists, 'You already set star on this discussion', {});
   }
 
   return output();
@@ -329,11 +343,11 @@ export async function removeDiscussionStar(r) {
     where: {
       discussionId: r.params.discussionId,
       userId: r.auth.credentials.id,
-    }
+    },
   });
 
   if (!starredDiscussion) {
-    return error(Errors.NotFound, "This discussion has no star", {});
+    return error(Errors.NotFound, 'This discussion has no star', {});
   }
 
   await starredDiscussion.destroy();
