@@ -1,12 +1,12 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import config from "../config/config";
+import config from '../config/config';
 import BigNumber from 'bignumber.js';
-import { output } from "../utils";
-import { ChainId, Token, TokenAmount, Pair } from "@pancakeswap/sdk";
-import { PancakeSwapApi } from "../controllers/controller.pancakeSwap";
-import { CoingeckoApi } from "../controllers/controller.coingecko";
-import { DailyLiquidity, WqtWbnbSwapEvent } from "@workquest/database-models/lib/models";
+import { output } from '../utils';
+import { ChainId, Token, TokenAmount, Pair } from '@pancakeswap/sdk';
+import { PancakeSwapApi } from '../controllers/controller.pancakeSwap';
+import { CoingeckoApi } from '../controllers/controller.coingecko';
+import { DailyLiquidity, WqtWbnbSwapEvent } from '@workquest/database-models/lib/models';
 
 const Web3 = require('web3');
 
@@ -18,33 +18,26 @@ const WQT = new Token(
   config.token.WQT.bscNetwork.name,
 );
 
-const WBNB = new Token(
-  ChainId.MAINNET,
-  config.token.WBNB.address,
-  config.token.WBNB.decimals,
-  config.token.WBNB.symbol,
-  config.token.WBNB.name,
-);
+const WBNB = new Token(ChainId.MAINNET, config.token.WBNB.address, config.token.WBNB.decimals, config.token.WBNB.symbol, config.token.WBNB.name);
 
-const pair = new Pair(
-  new TokenAmount(WQT, config.token.WQT.bscNetwork.amountMax),
-  new TokenAmount(WBNB, config.token.WBNB.amountMax)
-);
+const pair = new Pair(new TokenAmount(WQT, config.token.WQT.bscNetwork.amountMax), new TokenAmount(WBNB, config.token.WBNB.amountMax));
 
 const pancakeSwapApi = new PancakeSwapApi(pair);
 const coingeckoApi = new CoingeckoApi();
 
-const liquidityMiningProvider = new Web3( new Web3.providers.WebsocketProvider(config.contracts.liquidityMining.webSocketProvider, {
-  clientConfig: {
-    keepalive: true,
-    keepaliveInterval: 60000 // ms
-  },
-  reconnect: {
-    auto: true,
-    delay: 1000, // ms
-    onTimeout: false
-  }
-}));
+const liquidityMiningProvider = new Web3(
+  new Web3.providers.WebsocketProvider(config.contracts.liquidityMining.webSocketProvider, {
+    clientConfig: {
+      keepalive: true,
+      keepaliveInterval: 60000, // ms
+    },
+    reconnect: {
+      auto: true,
+      delay: 1000, // ms
+      onTimeout: false,
+    },
+  }),
+);
 
 const liquidityMiningAbiPath = path.join(__dirname, '../abi/WQLiquidityMining.json');
 const liquidityMiningAbi: [] = JSON.parse(fs.readFileSync(liquidityMiningAbiPath).toString()).abi;
@@ -54,7 +47,7 @@ export async function getSwaps(r) {
   const { count, rows } = await WqtWbnbSwapEvent.findAndCountAll({
     limit: r.query.limit,
     offset: r.query.offset,
-    order: [ ["timestamp", "DESC"] ],
+    order: [['timestamp', 'DESC']],
   });
 
   return output({ count, swaps: rows });
@@ -79,13 +72,13 @@ export async function getBurns(r) {
 }
 
 export async function getTokenDayData(r) {
-  const {count, rows} = await DailyLiquidity.findAndCountAll({
+  const { count, rows } = await DailyLiquidity.findAndCountAll({
     limit: r.query.limit,
     offset: r.query.offset,
-    order: [ ["date", "DESC"] ],
+    order: [['date', 'DESC']],
   });
 
-  return output({count, infoPer10Days: rows});
+  return output({ count, infoPer10Days: rows });
 }
 
 export async function getDistribution(r) {
@@ -95,7 +88,7 @@ export async function getDistribution(r) {
 
   const UsdPrice = await coingeckoApi.getUsdPrice();
 
-  const tokenDayData = await pancakeSwapApi.getTokenDayData({limit: 1, offset: 0});
+  const tokenDayData = await pancakeSwapApi.getTokenDayData({ limit: 1, offset: 0 });
 
   const reserveUSD = tokenDayData[0].reserveUSD;
   const totalSupply = tokenDayData[0].totalSupply;
@@ -105,6 +98,6 @@ export async function getDistribution(r) {
     priceUSD: UsdPrice.toString(),
     totalStaked: totalStaked.toString(),
     reserveUSD: reserveUSD,
-    totalSupply: totalSupply
+    totalSupply: totalSupply,
   });
 }
