@@ -1,8 +1,8 @@
-import { ChainId, Token, TokenAmount, Pair } from "@uniswap/sdk";
-import axios from "axios";
-import { error, output } from "../utils";
-import config from "../config/config";
-import { Errors } from "../utils/errors";
+import { ChainId, Token, TokenAmount, Pair } from '@uniswap/sdk';
+import axios from 'axios';
+import { error, output } from '../utils';
+import config from '../config/config';
+import { Errors } from '../utils/errors';
 
 const WQT = new Token(
   ChainId.MAINNET,
@@ -12,21 +12,12 @@ const WQT = new Token(
   config.token.WQT.ethereumNetwork.name,
 );
 
-const WETH = new Token(
-  ChainId.MAINNET,
-  config.token.WETH.address,
-  config.token.WETH.decimals,
-  config.token.WETH.symbol,
-  config.token.WETH.name,
-);
+const WETH = new Token(ChainId.MAINNET, config.token.WETH.address, config.token.WETH.decimals, config.token.WETH.symbol, config.token.WETH.name);
 
-const pair = new Pair(
-  new TokenAmount(WQT, config.token.WQT.ethereumNetwork.amountMax),
-  new TokenAmount(WETH, config.token.WETH.amountMax),
-);
+const pair = new Pair(new TokenAmount(WQT, config.token.WQT.ethereumNetwork.amountMax), new TokenAmount(WETH, config.token.WETH.amountMax));
 
 const api = axios.create({
-  baseURL: 'https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v2'
+  baseURL: 'https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v2',
 });
 
 export async function getSwaps(r) {
@@ -36,7 +27,7 @@ export async function getSwaps(r) {
         swaps(first:${r.query.limit}, skip:${r.query.offset}, orderBy: timestamp, orderDirection: desc, 
         where: { pair: "${pair.liquidityToken.address.toLowerCase()}" } ) 
         { transaction { id timestamp } pair { txCount }
-        amount0In amount0Out amount1In amount1Out amountUSD to } }`
+        amount0In amount0Out amount1In amount1Out amountUSD to } }`,
     });
 
     if (result.data.errors) {
@@ -51,11 +42,12 @@ export async function getSwaps(r) {
 
 export async function getMints(r) {
   try {
-    const result = await api.post('', {query: `{ 
+    const result = await api.post('', {
+      query: `{ 
       mints(first:${r.query.limit}, skip:${r.query.offset}, orderBy: timestamp, orderDirection: desc, 
       where: { pair: "${pair.liquidityToken.address.toLowerCase()}" }) 
       { transaction { id timestamp } pair { txCount}
-      to liquidity amount0 amount1 amountUSD } }`
+      to liquidity amount0 amount1 amountUSD } }`,
     });
 
     if (result.data.errors) {
@@ -75,7 +67,7 @@ export async function getBurns(r) {
         burns(first:${r.query.limit}, skip:${r.query.offset}, orderBy: timestamp, orderDirection: desc, 
         where: { pair: "${pair.liquidityToken.address.toLowerCase()}" }) 
         { transaction { id timestamp } pair { txCount }
-        to liquidity amount0 amount1 amountUSD } }`
+        to liquidity amount0 amount1 amountUSD } }`,
     });
 
     if (result.data.errors) {
@@ -97,7 +89,7 @@ export async function getTokenDayData(r) {
         where: {pairAddress: "${pair.liquidityToken.address.toLowerCase()}"})
         { date reserve0 reserve1 totalSupply reserveUSD dailyVolumeToken0
           dailyVolumeToken1 dailyVolumeUSD dailyTxns 
-        }}`
+        }}`,
     });
 
     if (result.data.errors) {
