@@ -11,8 +11,8 @@ function makeWhere(query) {
 }
 
 function makeOrderBy(sort) {
-  let order = ' ORDER BY ';
-  let byFields = [];
+  const order = ' ORDER BY ';
+  const byFields = [];
 
   for (const [column, sortBy] of Object.entries(sort)) {
     byFields.push(`"${column}" ${sortBy}`);
@@ -22,8 +22,8 @@ function makeOrderBy(sort) {
 }
 
 export async function mapPoints(r) {
-  let where = makeWhere(r);
-  let query = `
+  const where = makeWhere(r);
+  const query = `
   SELECT
        count as "pointsCount",
        id as "questId",
@@ -68,7 +68,7 @@ export async function mapPoints(r) {
     GROUP BY cid
     Order BY cid) LC;`;
 
-  const [results, ] = await r.server.app.db.query(query, {
+  const [results] = await r.server.app.db.query(query, {
     replacements: {
       northLongitude: r.query.northAndSouthCoordinates.north.longitude,
       northLatitude: r.query.northAndSouthCoordinates.north.latitude,
@@ -81,9 +81,9 @@ export async function mapPoints(r) {
 }
 
 export async function listMapPoints(r) {
-  let where = makeWhere(r);
-  let order = r.query.sort ? makeOrderBy(r.query.sort) : '';
-  let query = `
+  const where = makeWhere(r);
+  const order = r.query.sort ? makeOrderBy(r.query.sort) : '';
+  const query = `
   SELECT 
     id, "userId", status,
     priority, category, location,
@@ -93,7 +93,7 @@ export async function listMapPoints(r) {
   ${where}
   ${order}`;
 
-  const [results, ] = await r.server.app.db.query(query, {
+  const [results] = await r.server.app.db.query(query, {
     replacements: {
       northLongitude: r.query.northAndSouthCoordinates.north.longitude,
       northLatitude: r.query.northAndSouthCoordinates.north.latitude,
