@@ -1,27 +1,25 @@
-import {output } from '../utils';
-import {MediaController} from "../controllers/controller.media";
-import {UserController} from "../controllers/user/controller.user";
-import {PortfolioController} from "../controllers/user/controller.portfolio";
-import {
-  User,
-  UserRole,
-  Portfolio,
-} from "@workquest/database-models/lib/models";
+import { output } from '../utils';
+import { MediaController } from '../controllers/controller.media';
+import { UserController } from '../controllers/user/controller.user';
+import { PortfolioController } from '../controllers/user/controller.portfolio';
+import { User, UserRole, Portfolio } from '@workquest/database-models/lib/models';
 
 export async function addCase(r) {
   const medias = await MediaController.getMedias(r.payload.medias);
   const userController = new UserController(r.auth.credentials);
 
-  await userController
-    .userMustHaveRole(UserRole.Worker)
+  await userController.userMustHaveRole(UserRole.Worker);
 
   const transaction = await r.server.app.db.transaction();
 
-  const portfolioController = await PortfolioController.new({
-    userId: userController.user.id,
-    title: r.payload.title,
-    description: r.payload.description
-  }, transaction);
+  const portfolioController = await PortfolioController.new(
+    {
+      userId: userController.user.id,
+      title: r.payload.title,
+      description: r.payload.description,
+    },
+    transaction,
+  );
 
   await portfolioController.setMedias(medias, transaction);
 
@@ -37,17 +35,16 @@ export async function getCases(r) {
     where: { userId: workerController.user.id },
     limit: r.query.limit,
     offset: r.query.offset,
-    order: [ [ 'createdAt', 'DESC' ] ],
+    order: [['createdAt', 'DESC']],
   });
 
-  return output({count, cases: rows});
+  return output({ count, cases: rows });
 }
 
 export async function deleteCase(r) {
   const portfolioController = new PortfolioController(await Portfolio.findByPk(r.params.portfolioId));
 
-  portfolioController
-    .mustBeCaseCreator(r.auth.credentials.id)
+  portfolioController.mustBeCaseCreator(r.auth.credentials.id);
 
   await portfolioController.destroy();
 
@@ -60,8 +57,7 @@ export async function editCase(r) {
   const portfolio = await Portfolio.findByPk(r.params.portfolioId);
   const portfolioController = new PortfolioController(portfolio);
 
-  portfolioController
-    .mustBeCaseCreator(r.auth.credentials.id)
+  portfolioController.mustBeCaseCreator(r.auth.credentials.id);
 
   const transaction = await r.server.app.db.transaction();
 
@@ -70,7 +66,5 @@ export async function editCase(r) {
 
   await transaction.commit();
 
-  return output(
-    await Portfolio.findByPk(portfolio.id)
-  );
+  return output(await Portfolio.findByPk(portfolio.id));
 }
