@@ -1,58 +1,52 @@
-import {Op, literal} from "sequelize";
-import {error, output} from "../utils";
-import {Errors} from "../utils/errors";
-import {QuestController} from "../controllers/quest/controller.quest";
-import {
-  User,
-  Quest,
-  QuestStatus,
-  QuestDispute,
-  DisputeStatus, QuestChat
-} from "@workquest/database-models/lib/models";
+import { Op, literal } from 'sequelize';
+import { error, output } from '../utils';
+import { Errors } from '../utils/errors';
+import { QuestController } from '../controllers/quest/controller.quest';
+import { User, Quest, QuestStatus, QuestDispute, DisputeStatus, QuestChat } from '@workquest/database-models/lib/models';
 
 export async function openDispute(r) {
   const user: User = r.auth.credentials;
 
   const isDisputeExists = await QuestDispute.findOne({
-    where: { questId: r.params.questId }
+    where: { questId: r.params.questId },
   });
 
   if (isDisputeExists) {
-    return error(Errors.AlreadyExists,'Dispute for this quest already exists',{});
+    return error(Errors.AlreadyExists, 'Dispute for this quest already exists', {});
   }
 
   const quest = await Quest.findByPk(r.params.questId);
   const questController = new QuestController(quest);
 
-  questController
-    .userMustBelongToQuest(user.id)
-    .questMustHaveStatus(QuestStatus.Active, QuestStatus.WaitConfirm)
+  questController.userMustBelongToQuest(user.id).questMustHaveStatus(QuestStatus.Active, QuestStatus.WaitConfirm);
 
   // if (r.payload.reason === DisputeReason.poorlyDoneJob) {
   //   questController.questMustHaveStatus(QuestStatus.WaitConfirm);
   // }
 
-  const dayInMilliseconds = 86400000
-  const allowDate = quest.createdAt.getTime() + dayInMilliseconds;
+  const dayInMilliseconds = 86400000;
+  const allowDate = quest.startedAt.getTime() + dayInMilliseconds;
 
   if (allowDate > Date.now()) {
     return error(Errors.InvalidDate, 'Can open dispute after 24 hours after creating quest', {});
   }
 
-  const opponentUserId = quest.userId === user.id ?
-    quest.assignedWorkerId : quest.userId;
+  const opponentUserId = quest.userId === user.id ? quest.assignedWorkerId : quest.userId;
 
   const transaction = await r.server.app.db.transaction();
 
-  const dispute = await QuestDispute.create({
-    opponentUserId,
-    questId: quest.id,
-    openDisputeUserId: user.id,
-    status: DisputeStatus.pending,
-    openOnQuestStatus: quest.status,
-    reason: r.payload.reason,
-    problemDescription: r.payload.problemDescription,
-  }, { transaction });
+  const dispute = await QuestDispute.create(
+    {
+      opponentUserId,
+      questId: quest.id,
+      openDisputeUserId: user.id,
+      status: DisputeStatus.pending,
+      openOnQuestStatus: quest.status,
+      reason: r.payload.reason,
+      problemDescription: r.payload.problemDescription,
+    },
+    { transaction },
+  );
 
   await questController.openDispute(transaction);
 
@@ -63,18 +57,20 @@ export async function openDispute(r) {
 
 export async function getDispute(r) {
   const user: User = r.auth.credentials;
-  const questChatWorkerLiteral = literal(
-    '"quest->questChat"."workerId" = "quest"."assignedWorkerId"'
-  );
+  const questChatWorkerLiteral = literal('"quest->questChat"."workerId" = "quest"."assignedWorkerId"');
 
   const dispute = await QuestDispute.findByPk(r.params.disputeId, {
-    include: [{
-      model: Quest,
-      include: [{
-        model: QuestChat.unscoped(),
-        where: { questChatWorkerLiteral }
-      }]
-    }]
+    include: [
+      {
+        model: Quest,
+        include: [
+          {
+            model: QuestChat.unscoped(),
+            where: { questChatWorkerLiteral },
+          },
+        ],
+      },
+    ],
   });
 
   if (!dispute) {
@@ -83,8 +79,7 @@ export async function getDispute(r) {
 
   const questController = new QuestController(dispute.quest);
 
-  questController
-    .userMustBelongToQuest(user.id)
+  questController.userMustBelongToQuest(user.id);
 
   return output(dispute);
 }
@@ -92,10 +87,7 @@ export async function getDispute(r) {
 export async function getDisputes(r) {
   const { count, rows } = await QuestDispute.findAndCountAll({
     where: {
-      [Op.or]: [
-        { opponentUserId: r.auth.credentials.id },
-        { openDisputeUserId: r.auth.credentials.id },
-      ]
+      [Op.or]: [{ opponentUserId: r.auth.credentials.id }, { openDisputeUserId: r.auth.credentials.id }],
     },
     limit: r.query.limit,
     offset: r.query.offset,
@@ -104,5 +96,3 @@ export async function getDisputes(r) {
 
   return output({ count: count, disputes: rows });
 }
-
-
