@@ -1,5 +1,5 @@
-import { output } from "../utils";
-import {SkillsFiltersController} from "../controllers/controller.skillsFilters";
+import { output } from '../utils';
+import { SkillsFiltersController } from '../controllers/controller.skillsFilters';
 
 export async function getFilters(r) {
   const skillsFiltersController = await SkillsFiltersController.getInstance();
