@@ -1,9 +1,7 @@
 import { generateMediaHash, getUploadUrlS3 } from '../utils/storageService';
 import { output } from '../utils';
 import config from '../config/config';
-import {
-  Media
-} from "@workquest/database-models/lib/models";
+import { Media } from '@workquest/database-models/lib/models';
 
 export async function getUploadLink(r) {
   const hash = generateMediaHash(60);
