@@ -1,26 +1,24 @@
-import axios from "axios";
-import * as crypto from "crypto";
-import serverConfig from "../config/config";
-import { error, output } from "../utils";
-import * as FormData from "form-data";
-import { Errors } from "../utils/errors";
-import * as querystring from "querystring";
-import {
-  StatusKYC, User
-} from "@workquest/database-models/lib/models";
+import axios from 'axios';
+import * as crypto from 'crypto';
+import serverConfig from '../config/config';
+import { error, output } from '../utils';
+import * as FormData from 'form-data';
+import { Errors } from '../utils/errors';
+import * as querystring from 'querystring';
+import { StatusKYC, User } from '@workquest/database-models/lib/models';
 
 export enum ReviewAnswer {
-  Red = "RED",
-  Green = "GREEN",
+  Red = 'RED',
+  Green = 'GREEN',
 }
 
 const api = axios.create({
   baseURL: serverConfig.sumsub.baseURL,
   headers: {
-    "Accept": "application/json",
+    Accept: 'application/json',
     'Content-Type': 'application/json',
     'X-App-Token': serverConfig.sumsub.appToken,
-  }
+  },
 });
 
 api.interceptors.request.use(createSignature, function (error) {
@@ -28,12 +26,9 @@ api.interceptors.request.use(createSignature, function (error) {
 });
 
 function checkDigest(r): boolean {
-  const calculatedDigest = crypto
-    .createHmac('sha1', serverConfig.sumsub.secretKey)
-    .update(r.payload)
-    .digest('hex')
+  const calculatedDigest = crypto.createHmac('sha1', serverConfig.sumsub.secretKey).update(r.payload).digest('hex');
 
-  return calculatedDigest === r.headers['x-payload-digest']
+  return calculatedDigest === r.headers['x-payload-digest'];
 }
 
 function createSignature(config) {
@@ -56,13 +51,13 @@ function createSignature(config) {
 
 export async function createAccessToken(r) {
   if (r.auth.credentials.statusKYC === StatusKYC.Confirmed) {
-    return error(Errors.KYCAlreadyVerified, "User already verified", {});
+    return error(Errors.KYCAlreadyVerified, 'User already verified', {});
   }
 
   try {
     const qs = querystring.stringify({
       userId: r.auth.credentials.id,
-      ttlInSecs: serverConfig.sumsub.accessTokenTTL
+      ttlInSecs: serverConfig.sumsub.accessTokenTTL,
     });
     const result = await api.post(`/resources/accessTokens?` + qs);
 
@@ -78,12 +73,12 @@ export async function createAccessToken(r) {
 
 export async function applicantReviewed(r) {
   if (!checkDigest(r)) {
-    return error(Errors.InvalidPayload, "x-payload-digest failed", {});
+    return error(Errors.InvalidPayload, 'x-payload-digest failed', {});
   }
 
-  const payload = JSON.parse(r.payload.toString())
+  const payload = JSON.parse(r.payload.toString());
   if (payload.reviewResult.reviewAnswer !== ReviewAnswer.Green) {
-    return error(Errors.InvalidPayload, "Applicant not reviewed", {});
+    return error(Errors.InvalidPayload, 'Applicant not reviewed', {});
   }
 
   const user = await User.findOne({ where: { id: payload.externalUserId } });
