@@ -1,9 +1,9 @@
-import {config} from "dotenv";
+import { config } from 'dotenv';
 
-config({ path: __dirname +  '/../../../.env.bridge'});
+config({ path: __dirname + '/../../../.env.bridge' });
 
 export default {
-  debug: process.env.BRIDGE_DEBUG === "true",
+  debug: process.env.BRIDGE_DEBUG === 'true',
   ethereumMainNetwork: {
     parseEventsFromHeight: parseInt(process.env.BRIDGE_ETH_MAINNETWORK_PARSE_EVENTS_FROM_HEIGHT),
     contract: process.env.BRIDGE_ETH_MAINNETWORK_CONTRACT,
@@ -25,4 +25,4 @@ export default {
     webSocketProvider: process.env.BRIDGE_BSC_TESTNETWORK_WEBSOCKET_PROVIDER,
   },
   privateKey: process.env.BRIDGE_CONTRACT_PRIVAT_KEY,
-}
+};
