@@ -1,28 +1,30 @@
-import config from "./config";
+import config from './config';
 
 export default {
-  "pathPrefixSize": 2,
-  "basePath": "/api/",
-  "host": process.env.LOCAL ? "localhost:3000" : config.baseUrl.replace("https://", ""),
-  "grouping": "tags",
-  "schemes": ["https", "http"],
-  "info": {
-    "title": "WorkQuest Main Backend - API Documentation",
-    "version": "",
-    "description": "API Documentation\n\nYou can use https://mdenushev.github.io/nes-cli/ to test ws connections"
+  pathPrefixSize: 2,
+  basePath: '/api/',
+  host: process.env.LOCAL ? 'localhost:3000' : config.baseUrl.replace('https://', ''),
+  grouping: 'tags',
+  schemes: ['https', 'http'],
+  info: {
+    title: 'WorkQuest Main Backend - API Documentation',
+    version: '',
+    description: 'API Documentation\n\nYou can use https://mdenushev.github.io/nes-cli/ to test ws connections',
   },
-  "securityDefinitions": {
-    "Bearer": {
-      "type": "apiKey",
-      "name": "Authorization",
-      "in": "header",
-      "x-keyPrefix": "Bearer "
-    }
+  securityDefinitions: {
+    Bearer: {
+      type: 'apiKey',
+      name: 'Authorization',
+      in: 'header',
+      'x-keyPrefix': 'Bearer ',
+    },
   },
-  "security": [{
-    "Bearer": []
-  }],
-  "jsonPath": "/documentation.json",
-  "documentationPath": "/documentation",
-  "debug": true
+  security: [
+    {
+      Bearer: [],
+    },
+  ],
+  jsonPath: '/documentation.json',
+  documentationPath: '/documentation',
+  debug: true,
 };
