@@ -1,29 +1,29 @@
-import { Chat, ChatMember, ChatType, QuestChatStatuses, User } from "@workquest/database-models/lib/models";
-import { error } from "../../utils";
-import { Errors } from "../../utils/errors";
+import { Chat, ChatMember, ChatType, QuestChatStatuses, User } from '@workquest/database-models/lib/models';
+import { error } from '../../utils';
+import { Errors } from '../../utils/errors';
 
 abstract class ChatHelper {
   public abstract chat: Chat;
 
   static async chatMustExists(chatId: string) {
-    if (!await Chat.findByPk(chatId)) {
-      throw error(Errors.NotFound, "Chat does not exist", { chatId });
+    if (!(await Chat.findByPk(chatId))) {
+      throw error(Errors.NotFound, 'Chat does not exist', { chatId });
     }
   }
 
   public async chatMustHaveMember(userId: string) {
     const member = await ChatMember.findOne({
-      where: { chatId: this.chat.id, userId }
+      where: { chatId: this.chat.id, userId },
     });
 
     if (!member) {
-      throw error(Errors.Forbidden, "User is not a member of this chat", {});
+      throw error(Errors.Forbidden, 'User is not a member of this chat', {});
     }
   }
 
   public chatMustHaveType(type: ChatType): this {
     if (this.chat.type !== type) {
-      throw error(Errors.InvalidType, "Type does not match", {});
+      throw error(Errors.InvalidType, 'Type does not match', {});
     }
 
     return this;
@@ -31,9 +31,9 @@ abstract class ChatHelper {
 
   public questChatMastHaveStatus(status: QuestChatStatuses): this {
     if (this.chat.questChat.status !== status) {
-      throw error(Errors.Forbidden, "Quest chat type does not match", {
+      throw error(Errors.Forbidden, 'Quest chat type does not match', {
         mastHave: status,
-        current: this.chat.questChat.status
+        current: this.chat.questChat.status,
       });
     }
 
@@ -42,7 +42,7 @@ abstract class ChatHelper {
 
   public chatMustHaveOwner(userId: string): this {
     if (this.chat.ownerUserId !== userId) {
-      throw error(Errors.Forbidden, "User is not a owner in this chat", {});
+      throw error(Errors.Forbidden, 'User is not a owner in this chat', {});
     }
 
     return this;
@@ -54,9 +54,7 @@ abstract class ChatHelper {
     });
 
     if (members.length !== 0) {
-      const existsIds = userIds.filter(userId =>
-        members.findIndex(member => userId === member.userId) !== -1
-      );
+      const existsIds = userIds.filter((userId) => members.findIndex((member) => userId === member.userId) !== -1);
 
       throw error(Errors.AlreadyExists, 'Users already exists in group chat', { existsIds });
     }
@@ -66,14 +64,11 @@ abstract class ChatHelper {
 }
 
 export class ChatController extends ChatHelper {
-  constructor(
-    public chat: Chat
-  ) {
+  constructor(public chat: Chat) {
     super();
 
     if (!chat) {
-      throw error(Errors.NotFound, "Chat not found", {});
+      throw error(Errors.NotFound, 'Chat not found', {});
     }
   }
-
 }
