@@ -1,12 +1,12 @@
-import { Message } from "@workquest/database-models/lib/models";
-import { error } from "../../utils";
-import { Errors } from "../../utils/errors";
-import { Transaction } from "sequelize";
+import { Message } from '@workquest/database-models/lib/models';
+import { error } from '../../utils';
+import { Errors } from '../../utils/errors';
+import { Transaction } from 'sequelize';
 
 abstract class MessageHelper {
   public abstract message: Message;
 
-  async messageMustBeSender(userId: String) {
+  async messageMustBeSender(userId: string) {
     if (this.message.senderUserId !== userId) {
       throw error(Errors.Forbidden, "User isn't sender of the message", {
         messageId: this.message.id,
@@ -16,9 +16,9 @@ abstract class MessageHelper {
     return this;
   }
 
-  async messageMustBeChat(chatId: String) {
+  async messageMustBeChat(chatId: string) {
     if (this.message.chatId !== chatId) {
-      throw error(Errors.Forbidden, "This message not from this chat", {});
+      throw error(Errors.Forbidden, 'This message not from this chat', {});
     }
 
     return this;
@@ -26,15 +26,11 @@ abstract class MessageHelper {
 }
 
 export class MessageController extends MessageHelper {
-
-  constructor(
-    public message: Message
-  ) {
+  constructor(public message: Message) {
     super();
 
     if (!message) {
-      throw error(Errors.NotFound, "Message not found", {});
+      throw error(Errors.NotFound, 'Message not found', {});
     }
   }
-
 }
