@@ -1,16 +1,16 @@
-import { Media } from "@workquest/database-models/lib/models";
-import { error } from "../utils";
-import { Errors } from "../utils/errors";
-import { isMediaExists } from "../utils/storageService";
+import { Media } from '@workquest/database-models/lib/models';
+import { error } from '../utils';
+import { Errors } from '../utils/errors';
+import { isMediaExists } from '../utils/storageService';
 
 export class MediaController {
   static async getMedia(mediaId: string): Promise<Media> {
     const media = await Media.findByPk(mediaId);
 
     if (!media) {
-      throw error(Errors.NotFound, 'Media is not found', { mediaId })
+      throw error(Errors.NotFound, 'Media is not found', { mediaId });
     }
-    if (!await isMediaExists(media)) {
+    if (!(await isMediaExists(media))) {
       throw error(Errors.InvalidPayload, 'Media is not exists', { mediaId: media.id });
     }
 
@@ -18,7 +18,7 @@ export class MediaController {
   }
   static async getMedias(mediaIds: string[]): Promise<Media[]> {
     const medias = await Media.findAll({
-      where: { id: mediaIds }
+      where: { id: mediaIds },
     });
 
     // if (medias.length !== mediaIds.length) {
