@@ -1,18 +1,13 @@
-import {error} from "../../utils";
-import { Op, Transaction } from "sequelize";
-import {Errors} from "../../utils/errors";
-import {
-  Quest,
-  QuestsResponse,
-  QuestsResponseStatus,
-  QuestsResponseType, User
-} from "@workquest/database-models/lib/models";
+import { error } from '../../utils';
+import { Op, Transaction } from 'sequelize';
+import { Errors } from '../../utils/errors';
+import { Quest, QuestsResponse, QuestsResponseStatus, QuestsResponseType, User } from '@workquest/database-models/lib/models';
 
 abstract class QuestsResponseHelper {
   public abstract questsResponse: QuestsResponse;
 
   /** Checks list */
-  public workerMustBeInvitedToQuest(workerId: String): QuestsResponseHelper {
+  public workerMustBeInvitedToQuest(workerId: string): QuestsResponseHelper {
     this.questsResponseMustHaveType(QuestsResponseType.Invite);
 
     if (this.questsResponse.workerId !== workerId) {
@@ -54,26 +49,28 @@ abstract class QuestsResponseHelper {
 }
 
 export class QuestsResponseController extends QuestsResponseHelper {
-
-  constructor(
-    public questsResponse: QuestsResponse
-  ) {
+  constructor(public questsResponse: QuestsResponse) {
     super();
 
     if (!questsResponse) {
-      throw error(Errors.NotFound, "QuestsResponse not found", {});
+      throw error(Errors.NotFound, 'QuestsResponse not found', {});
     }
   }
 
   async closeOtherResponsesToQuest(quest: Quest, transaction?: Transaction) {
     try {
-      await QuestsResponse.update({
-        status: QuestsResponseStatus.Closed
-      }, {
-      where: {
-        id: { [Op.ne]: this.questsResponse.id },
-        questId: quest.id,
-      }, transaction });
+      await QuestsResponse.update(
+        {
+          status: QuestsResponseStatus.Closed,
+        },
+        {
+          where: {
+            id: { [Op.ne]: this.questsResponse.id },
+            questId: quest.id,
+          },
+          transaction,
+        },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -84,8 +81,13 @@ export class QuestsResponseController extends QuestsResponseHelper {
 
   static async closeAllResponsesOnQuest(quest: Quest, transaction?: Transaction) {
     try {
-      await QuestsResponse.update({ status: QuestsResponseStatus.Closed }, {
-        where: { questId: quest.id }, transaction });
+      await QuestsResponse.update(
+        { status: QuestsResponseStatus.Closed },
+        {
+          where: { questId: quest.id },
+          transaction,
+        },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -108,23 +110,31 @@ export class QuestsResponseController extends QuestsResponseHelper {
 
   static async reopenQuestResponses(quest: Quest, rejectedWorker: User, transaction?: Transaction) {
     try {
-      await QuestsResponse.update({
-        status: QuestsResponseStatus.Open,
-      },{
-        where: {
-          questId: quest.id,
-          workerId: { [Op.ne]: rejectedWorker.id }
-        }, transaction,
-      });
-      await QuestsResponse.update({
-        status: QuestsResponseStatus.Rejected,
-        previousStatus: QuestsResponseStatus.Rejected,
-      }, {
-        where: {
-          questId: quest.id,
-          workerId: rejectedWorker.id
-        }, transaction,
-      });
+      await QuestsResponse.update(
+        {
+          status: QuestsResponseStatus.Open,
+        },
+        {
+          where: {
+            questId: quest.id,
+            workerId: { [Op.ne]: rejectedWorker.id },
+          },
+          transaction,
+        },
+      );
+      await QuestsResponse.update(
+        {
+          status: QuestsResponseStatus.Rejected,
+          previousStatus: QuestsResponseStatus.Rejected,
+        },
+        {
+          where: {
+            questId: quest.id,
+            workerId: rejectedWorker.id,
+          },
+          transaction,
+        },
+      );
     } catch (e) {
       if (transaction) {
         await transaction.rollback();
@@ -133,4 +143,3 @@ export class QuestsResponseController extends QuestsResponseHelper {
     }
   }
 }
-
