@@ -1,14 +1,14 @@
-import {addJob} from "../utils/scheduler";
-import {ChatMember} from "@workquest/database-models/lib/models";
-import {Op} from "sequelize"
+import { addJob } from '../utils/scheduler';
+import { ChatMember } from '@workquest/database-models/lib/models';
+import { Op } from 'sequelize';
 
 export type UnreadMessageIncrementPayload = {
   chatId: string;
   notifierUserId?: string;
-}
+};
 
 export async function incrementUnreadCountMessageOfMembersJob(payload: UnreadMessageIncrementPayload) {
-  return addJob("incrementUnreadCountMessageOfMembers", payload);
+  return addJob('incrementUnreadCountMessageOfMembers', payload);
 }
 
 export default async function incrementUnreadCountMessageOfMembers(payload: UnreadMessageIncrementPayload) {
@@ -16,7 +16,6 @@ export default async function incrementUnreadCountMessageOfMembers(payload: Unre
     where: {
       ...(payload.notifierUserId && { userId: { [Op.ne]: payload.notifierUserId } }),
       chatId: payload.chatId,
-    }
+    },
   });
 }
-
