@@ -1,13 +1,13 @@
-import { addJob } from "../utils/scheduler";
-import { ChatMember, ChatsStatistic } from "@workquest/database-models/lib/models";
-import { Op } from "sequelize";
+import { addJob } from '../utils/scheduler';
+import { ChatMember, ChatsStatistic } from '@workquest/database-models/lib/models';
+import { Op } from 'sequelize';
 
 export type UserUnreadChatsPayload = {
-  userIds: string[],
-}
+  userIds: string[];
+};
 
 export async function updateCountUnreadChatsJob(payload: UserUnreadChatsPayload) {
-  return addJob("updateCountUnreadChats", payload);
+  return addJob('updateCountUnreadChats', payload);
 }
 
 export default async function updateCountUnreadChats(payload: UserUnreadChatsPayload) {
@@ -20,7 +20,8 @@ export default async function updateCountUnreadChats(payload: UserUnreadChatsPay
     });
 
     const [chatsStatistic, isCreated] = await ChatsStatistic.findOrCreate({
-      where: { userId }, defaults: { userId, unreadCountChats: unreadChatsCounter }
+      where: { userId },
+      defaults: { userId, unreadCountChats: unreadChatsCounter },
     });
 
     if (!isCreated) {
@@ -28,4 +29,3 @@ export default async function updateCountUnreadChats(payload: UserUnreadChatsPay
     }
   }
 }
-
