@@ -1,11 +1,5 @@
-import { addJob } from "../utils/scheduler";
-import {
-  Quest,
-  UserRole,
-  QuestStatus,
-  QuestsStatistic,
-  activeFlowStatuses,
-} from "@workquest/database-models/lib/models";
+import { addJob } from '../utils/scheduler';
+import { Quest, UserRole, QuestStatus, QuestsStatistic, activeFlowStatuses } from '@workquest/database-models/lib/models';
 
 export interface Data {
   userId: string;
@@ -15,10 +9,10 @@ export interface Data {
 type Statistic = {
   opened: number;
   completed: number;
-}
+};
 
 export async function updateQuestsStatisticJob(payload: Data) {
-  return addJob("updateQuestsStatistic", payload);
+  return addJob('updateQuestsStatistic', payload);
 }
 
 async function getWorkerQuestStatistic(workerId: string): Promise<Statistic> {
@@ -26,20 +20,18 @@ async function getWorkerQuestStatistic(workerId: string): Promise<Statistic> {
     where: {
       assignedWorkerId: workerId,
       status: QuestStatus.Done,
-    }
+    },
   });
   const openedQuestsPromises = Quest.unscoped().count({
     where: {
       assignedWorkerId: workerId,
       status: activeFlowStatuses,
-    }
+    },
   });
 
-  const [completed, opened] = await Promise.all([
-    completedQuestsPromises, openedQuestsPromises,
-  ]);
+  const [completed, opened] = await Promise.all([completedQuestsPromises, openedQuestsPromises]);
 
-  return { opened, completed }
+  return { opened, completed };
 }
 
 async function getEmployerQuestStatistic(employerId: string): Promise<Statistic> {
@@ -47,31 +39,28 @@ async function getEmployerQuestStatistic(employerId: string): Promise<Statistic>
     where: {
       userId: employerId,
       status: QuestStatus.Done,
-    }
+    },
   });
   const openedQuestsPromises = Quest.unscoped().count({
     where: {
       userId: employerId,
       status: activeFlowStatuses,
-    }
+    },
   });
 
-  const [completed, opened] = await Promise.all([
-    completedQuestsPromises, openedQuestsPromises,
-  ]);
+  const [completed, opened] = await Promise.all([completedQuestsPromises, openedQuestsPromises]);
 
-  return { opened, completed }
+  return { opened, completed };
 }
 
 export default async function updateQuestsStatistic(payload: Data) {
-  const [questsStatistic, ] = await QuestsStatistic.findOrCreate({
-    where:  { userId: payload.userId },
+  const [questsStatistic] = await QuestsStatistic.findOrCreate({
+    where: { userId: payload.userId },
     defaults: { userId: payload.userId },
   });
 
-  const statistic = payload.role === UserRole.Worker ?
-    await getWorkerQuestStatistic(payload.userId) : await getEmployerQuestStatistic(payload.userId);
+  const statistic =
+    payload.role === UserRole.Worker ? await getWorkerQuestStatistic(payload.userId) : await getEmployerQuestStatistic(payload.userId);
 
   await questsStatistic.update(statistic);
 }
-
