@@ -1,5 +1,3 @@
-import v1 from "./v1";
+import v1 from './v1';
 
-export default [
-  ...v1
-]
+export default [...v1];
