@@ -1,4 +1,4 @@
-import * as Joi from "joi";
+import * as Joi from 'joi';
 import * as handlers from '../../api/quest';
 import {
   outputOkSchema,
