@@ -1,24 +1,24 @@
-import { Pool } from "pg";
-import config from "../config/config";
+import { Pool } from 'pg';
+import config from '../config/config';
 
 // TODO: Maybe should rewrite and document function
 export async function addJob(taskName: string, payload?: any, userOptions: any = {}) {
   try {
     const pool = new Pool({ connectionString: config.dbLink });
-    let args = ["$1::text"];
+    const args = ['$1::text'];
     const options = { ...{ max_attempts: 25 }, ...userOptions };
     const values = [taskName];
     if (payload) {
-      args.push("payload := $2");
+      args.push('payload := $2');
       values.push(JSON.stringify(payload));
     }
 
     if (options) {
       let argumentPos = values.length;
-      ["queue_name", "max_attempts", "run_at", "interval"].forEach((col) => {
+      ['queue_name', 'max_attempts', 'run_at', 'interval'].forEach((col) => {
         if (options[col]) {
           argumentPos++;
-          if (col === "interval") {
+          if (col === 'interval') {
             args.push(`run_at := NOW() + ($${argumentPos} * INTERVAL '1 minute')`);
           } else {
             args.push(`${col} := $${argumentPos}`);
@@ -28,7 +28,7 @@ export async function addJob(taskName: string, payload?: any, userOptions: any =
         }
       });
     }
-    let query = `SELECT graphile_worker.add_job(${args.join(", ")});`;
+    const query = `SELECT graphile_worker.add_job(${args.join(', ')});`;
     await pool.query(query, values);
     await pool.end();
   } catch (e) {
